requirements.txt
flask
tensorflow
numpy
opencv-python
fpdf
Pillow
